requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.9
idna==2.7
Jinja2==2.10
livereload==2.5.2
Markdown==3.0
MarkupSafe==1.0
mkdocs==1.0.4
mock-open==1.3.1
more-itertools==4.3.0
pkginfo==1.4.2
pluggy==0.7.1
py==1.6.0
pytest==3.8.1
PyYAML==3.13
requests==2.19.1
requests-toolbelt==0.8.0
six==1.11.0
tornado==5.1.1
tqdm==4.26.0
twine==1.11.0
urllib3==1.23
